converted double data type of reviewScore and handling corousel button

+++ libraries/cm_push_notifications/src/main/java/com/tokopedia/notifications/model/ProductInfo.kt
@@ -83,7 +83,7 @@ data class ProductInfo(
         @SerializedName(CMConstant.PayloadKeys.REVIEW_SCORE)
         @ColumnInfo(name = CMConstant.PayloadKeys.REVIEW_SCORE)
         @Expose
-        var reviewScore: Long? = 0L,
+        var reviewScore: Double? = 0.0,
 
         @SerializedName(CMConstant.PayloadKeys.REVIEW_NUMBER)
         @ColumnInfo(name = CMConstant.PayloadKeys.REVIEW_NUMBER)
